Magnifying glass effect added to mouse position in place of fish eye effect
I couldn't figure out how to get the fish eye effect to keep the entire level in screen so I opted for a magnifying glasses effect instead. This is currently implemented on mouse position. Ideally this will be centred on the player character, then the player can unsnap the magnifying glass so investigate the whole level and snap it back to the player character when ready to keep moving